+++ packages/manager/src/features/Linodes/LinodesDetail/VolumesUpgradeBanner.tsx
@@ -7,35 +7,13 @@ import { Notice } from 'src/components/Notice/Notice';
 import { Paper } from 'src/components/Paper';
 import { Stack } from 'src/components/Stack';
 import { Typography } from 'src/components/Typography';
+import { getUpgradeableVolumeIds } from 'src/features/Volumes/utils';
 import { useNotificationsQuery } from 'src/queries/accountNotifications';
 import { useLinodeVolumesQuery } from 'src/queries/volumes';
-
-import type { Notification, Volume } from '@linode/api-v4';
 
 interface Props {
   linodeId: number;
 }
-
-export const getUpgradeableVolumeIds = (
-  volumes: Volume[],
-  notifications: Notification[]
-) => {
-  const upgradeableVolumeIds: number[] = [];
-
-  for (const volume of volumes) {
-    if (
-      notifications?.some(
-        (notification) =>
-          notification.entity?.id === volume.id &&
-          notification.type === 'volume_migration_scheduled'
-      )
-    ) {
-      upgradeableVolumeIds.push(volume.id);
-    }
-  }
-
-  return upgradeableVolumeIds;
-};
 
 export const VolumesUpgradeBanner = ({ linodeId }: Props) => {
   const history = useHistory();
